dosen semua

+++ data-dosen.php
@@ -44,8 +44,9 @@
             
         echo "<td>".$row['nama']."</td>";
         echo "<td>".$row['npk']."</td>" ; 
+        
         echo "<td><a href='edit-dosen.php?npk=".$row['npk']."'>Edit</a></td>";
-        echo"<td><a href='delete-dosen.php?npk=".$row['npk']."'>Delete</a></td>";
+        echo "<td><a href='delete-dosen.php?npk=".$row['npk']."' onclick='return confirm(\"Yakin ingin menghapus dosen ini?\");'>Delete</a></td>";
         echo "<input type='hidden' name='npk_lama' value='".$row['npk']."'>";
         echo"</tr>";
         }
